changed folder name for service to api in frontend under lib folder, also renamed files to match foldername
