+++ backend/routers/recipelistsRouter.js
@@ -31,7 +31,7 @@ router.get('/api/recipelists/user/:userId', async (req, res) => {
       include: {
         recipes: {
           include: {
-            categories: true,
+            category: true,
           },
         },
       },
