+++ frontend/src/lib/components/category-select/category-select.svelte
@@ -1,6 +1,6 @@
 <script>
   import * as Select from "$lib/components/ui/select/index.js";
-  import { getRecipeListsByUserId } from "$lib/services/recipelistService.js";
+  import { getRecipeListsByUserId } from "$lib/api/recipelistApi.js";
   import { Check } from "lucide-svelte";
   import { createEventDispatcher, onMount } from "svelte";
