+++ frontend/src/lib/components/delete-recipe-dialog/delete-recipe-dialog.svelte
@@ -3,7 +3,7 @@
   import { Button, buttonVariants } from "$lib/components/ui/button/index.js";
   import { X } from "lucide-svelte";
   import AlertDialogAction from "../ui/alert-dialog/alert-dialog-action.svelte";
-  import { deleteRecipe } from "$lib/services/recipeService";
+  import { deleteRecipe } from "$lib/api/recipeApi";
   import { toast } from "svelte-sonner";
 
   let { recipeId, selectedList = $bindable() } = $props();
