+++ frontend/src/routes/login/reset-password/[token]/+page.svelte
@@ -2,7 +2,7 @@
     import { Input } from "$lib/components/ui/input/index.js";
     import { Label } from "$lib/components/ui/label/index.js";
     import { Button } from "$lib/components/ui/button/index.js";
-    import { authService } from "$lib/services/authService";
+    import { authService } from "$lib/api/authApi";
     import * as Card from "$lib/components/ui/card";
     import { Toaster } from "$lib/components/ui/sonner";
     import { toast } from "svelte-sonner";
